Added a new hiding module for hiding values, shape and dimensionality of arrays. Below is some slightly updated text from an email I sent proposing the change:
Some of the benchmark kernels we are using with SaC require static inputs in the
code but require some information about these inputs to be hidden from the code
for the kernel being benchmarked. For example sometimes we want to hide the
length of a vector input so that SaC cannot specialise the code for vectors of
that length. The common technique to do this appears to be to write a special id
function that returns its input but with a less specialist type. These id
functions exploit the fact that SaC does not optimise between non-inlined
functions. The danger when doing this is that these functions may not be future
proof, for example if an optimisation is added to inline functions that are only
called once then this trick will no longer work.

I propose to add a directory to the standard library  and in there to add a
module called hiding. In the hiding module we can store functions for hiding
type information of static inputs for benchmarking. If benchmark implementations
make use of this library then providing the libraries stay maintained the
benchmark implementations will not become invalid after new compiler
optimisations are added.