Resolve target instead of library as dependencies
With this commit, we have moved from resolving intermodule dependencies
though library file paths to using targets instead. This resolves an
issue with builds whereby modules were being rebuild continually. The
reason for this was a circular dependency on both the lib*Mod and
lib*Tree files of the same module for various targets - these files
would be seen as outdated during the build, resulting in them being
rebuild, again and again.