Change Math.sac externals to use `#pragma header'
When compiling with NVCC, all common math operations and functions are
injected into the code. When SAC generates the C/CU code, it adds
prototypes for the external declarations. This leads to a conflict with
NVCC linker being unable to disambiguate which prototype is the right
one to use.

The change/fix is to use `#pragma header' and explicitly specify that we
want to use the prototype provided by the NVCC (or even the system).