fix NVCC errors (C++ related)
When we compile with the cuda target, we use NVCC to compile both SaC
and C code. NVCC treats all the source code as C++. There are some parts of
the C sources within the stdlib which make uses of features that do not
exist within C++.

For instance the `restrict` pointer attribute, exists as a keyword
in C99, but not as part of C++. For GCC and CLANG for C/C++ there is
defined the `__restrict__` attribute which serves the same purpose as
`restrict`.

This commit makes the appropriate changes, now NVCC is happy.