generate documentation using Doxygen
This commit adds support to generate documention using the Doxygen tool.
Principly this is a straightforward task, however the SaC language has
certain properties which do not work well with Doxygen's adherences to
C/C++ syntax conventions.

Through the included script, sac2doxygen.py, we are able to transform
the SaC code into a form that is more suitable for Doxygen.

The script either converts or removes certain keywords and statements.
Additionally it runs CPP beforehand, expanding all macros - this gives
us all function definitions which is more useful when referencing the
documentation.