fix missing FiberIO header file dependency
When building C files, the use of `IMPLICIT DEPENDS` from CMake does not
seem to correctly track _generated_ C header files. At initial
compilation, using `make -j` for instance, this causes the build to
reach a failed state due to header file `FibreScan.tab.h` not being
found.

This header file is generated, and at the level of SaC module files, is
correctly tracked as there is an explicit dependency. No so in the case
for C files, where we only use `IMPLICIT DEPENDS`.

This commit fixes this by performing an additional check that explicitly
addes the `FibreScan.tab.h` file as a dependency for the given C
compilation unit.